Fix Datastore abstraction
- Only AppEngine Datastore should support memcache
  - hence `LoadTestRunsByKeys()` depends on Datastore implementation
    - hence update its call-sites
- Only AppEngine Datastore behaves as a store for `CachedStore`
  - hence have it (and not other Datastore impls) implement it directly
