+++ util/populate_dev_data.go
@@ -275,7 +275,7 @@ func copyProdRuns(ctx context.Context, filters shared.TestRunFilter) {
 			}
 			if len(shas) > 0 {
 				sha = shas[0]
-				if loaded, err := shared.LoadTestRunsByKeys(store, keys[sha]); err != nil {
+				if loaded, err := store.LoadTestRunsByKeys(keys[sha]); err != nil {
 					log.Printf("Failed to load test runs by keys: %s", err.Error())
 					continue
 				} else {
